Move ovssubnet to pkg dir
