+++ pkg/ovssubnet/controller/kube/bin/openshift-sdn-kube-subnet-setup.sh
